requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
python-multipart==0.0.22
pydantic==2.12.0
deepface==0.0.99
easyocr==1.7.2
opencv-python-headless==4.13.0.92
Pillow==12.1.1
numpy==2.4.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
tf-keras==2.20.1
